+++ md5_phone_store_management/src/main/java/com/example/md5_phone_store_management/service/ITransactionOutService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface ITransactionOutService {
     List<InventoryTransaction> getAllOutTransactions();
     List<InventoryTransaction> getAllInTransactions();
-    Optional<InventoryTransaction> getOutTransactionById(int id);
+    Optional<InventoryTransaction> getOutTransactionById(Long id);
     void addOutTransaction(InventoryTransaction transaction);
     void updateOutTransaction(int id, InventoryTransaction transaction);
     void deleteOutTransaction(int id);
